Restructured BLE service
DeviceScanActivity now properly only scan for devices (as well as verifying permissions and enabling bluetooth). I added a BLE service that manage everything related to BLE. It's therefore easier to understand the program (at least, I think so). DeviceActivity at the moment connect, and on connect the services and characteristics are displayed in the log as debug. :smile: