Disconnection fix
When the button 'disconnect' is called on the DeviceActivity layout, the disconnect could cause some problems.  Now proper disconnection is verified by seeing the ACTION_DISCONNECTED broadcasted from the service.  On the getBattery method I also added a verification that mGatt isn't null to avoid invoke on a null object reference.